\name{intervals}
\alias{intervals}
\title{Calculate intervals}
\usage{
  intervals(mobility, minimum = 0, smooth = TRUE)
}
\arguments{
  \item{mobility}{mobility vector}

  \item{minimum}{minimal interval span.}
}
\value{
  data frame
}
\description{
  The function calculates the number of intervals of
  identical mobility modes
}

