\name{painreport}
\alias{painreport}
\title{Generate pain pilot report}
\usage{
  painreport(serverurl, token, username, days = 14)
}
\arguments{
  \item{serverurl}{url of the ohmage server}

  \item{token}{valid auth token}

  \item{username}{target user}

  \item{days}{number of days}
}
\description{
  Generate pain pilot report
}

